+++ resources/views/AJ/edit5_4.blade.php
@@ -151,7 +151,17 @@
         }
       var formData = new FormData(this);
 
-      $.ajax({
+      
+
+      swal({
+      title: "ยืนยันการบันทึก?",
+      icon: "warning",
+      buttons: true,
+      successMode: true,
+    })
+    .then((willDelete) => {
+      if (willDelete) {
+        $.ajax({
         type: 'POST',
         url: "/updateindicator5_4",
         data: formData,
@@ -161,7 +171,7 @@
         dataType: 'json',
         success: (data) => {
           swal({
-          title: "แก้ไขข้อมูลเรียบร้อยแล้ว",
+          title: "แก้ไขข้อมูลเรียบร้อย",
           text: "",
           icon: "success",
           button: "ตกลง",
@@ -174,10 +184,24 @@
           console.log(data.responseJSON.errors);
         }
       });
+      } else {
+        
+      }
+    });
     });
     $('.delete').click(function(e) {
       var id = $(this).attr('id');
-      $.ajax({
+      
+
+      swal({
+      title: "ยืนยันการลบข้อมูล?",
+      icon: "warning",
+      buttons: true,
+      successMode: true,
+    })
+    .then((willDelete) => {
+      if (willDelete) {
+        $.ajax({
         type: 'delete',
         url: "/deletedoc5_4/"+id,
         data: {id:id},
@@ -187,7 +211,7 @@
         dataType: 'json',
         success: (data) => {
           swal({
-          title: "ลบข้อมูลเรียบร้อยแล้ว",
+          title: "ลบข้อมูลเรียบร้อย",
           text: "",
           icon: "success",
           button: "ตกลง",
@@ -200,6 +224,10 @@
           console.log(data.responseJSON.errors);
         }
       });
+      } else {
+        
+      }
+    });
     });
   });
