+++ resources/views/AJ/editA.blade.php
@@ -154,7 +154,17 @@
         formData.append('files' + i, files.files[i]);
       }
       formData.append('TotalFiles', TotalFiles);
-      $.ajax({
+      
+
+      swal({
+      title: "ยืนยันการบันทึก?",
+      icon: "warning",
+      buttons: true,
+      successMode: true,
+    })
+    .then((willDelete) => {
+      if (willDelete) {
+        $.ajax({
         type: 'POST',
         url: "/updatea",
         data: formData,
@@ -164,7 +174,7 @@
         dataType: 'json',
         success: (data) => {
           swal({
-          title: "บันทึกข้อมูลสำเร็จ",
+          title: "แก้ไขข้อมูลเรียบร้อย",
           text: "",
           icon: "success",
           button: "ตกลง",
@@ -177,11 +187,25 @@
           console.log(data.responseJSON.errors);
         }
       });
+      } else {
+        
+      }
+    });
     });
 
     $('.delete').click(function(e) {
       var id = $(this).attr('id');
-      $.ajax({
+      
+
+      swal({
+      title: "ยืนยันการลบข้อมูล?",
+      icon: "warning",
+      buttons: true,
+      successMode: true,
+    })
+    .then((willDelete) => {
+      if (willDelete) {
+        $.ajax({
         type: 'delete',
         url: "/deletedoc/"+id,
         data: {id:id},
@@ -191,7 +215,7 @@
         dataType: 'json',
         success: (data) => {
           swal({
-          title: "ลบข้อมูลเรียบร้อยแล้ว",
+          title: "ลบข้อมูลเรียบร้อย",
           text: "",
           icon: "success",
           button: "ตกลง",
@@ -204,6 +228,10 @@
           console.log(data.responseJSON.errors);
         }
       });
+      } else {
+        
+      }
+    });
     });
   });
 </script>
