+++ resources/views/AJ/editp.blade.php
@@ -155,7 +155,17 @@
         formData.append('files' + i, files.files[i]);
       }
       formData.append('TotalFiles', TotalFiles);
-      $.ajax({
+      
+
+      swal({
+      title: "ยืนยันการบันทึก?",
+      icon: "warning",
+      buttons: true,
+      successMode: true,
+    })
+    .then((willDelete) => {
+      if (willDelete) {
+        $.ajax({
         type: 'POST',
         url: "/updatep",
         data: formData,
@@ -165,7 +175,7 @@
         dataType: 'json',
         success: (data) => {
           swal({
-          title: "บันทึกข้อมูลสำเร็จ",
+          title: "แก้ไขข้อมูลเรียบร้อย",
           text: "",
           icon: "success",
           button: "ตกลง",
@@ -178,11 +188,25 @@
           console.log(data.responseJSON.errors);
         }
       });
+      } else {
+        
+      }
+    });
     });
 
     $('.delete').click(function(e) {
       var id = $(this).attr('id');
-      $.ajax({
+      
+
+      swal({
+      title: "ยืนยันการลบข้อมูล?",
+      icon: "warning",
+      buttons: true,
+      successMode: true,
+    })
+    .then((willDelete) => {
+      if (willDelete) {
+        $.ajax({
         type: 'delete',
         url: "/deletedoc/"+id,
         data: {id:id},
@@ -192,7 +216,7 @@
         dataType: 'json',
         success: (data) => {
           swal({
-          title: "ลบข้อมูลเรียบร้อยแล้ว",
+          title: "ลบข้อมูลเรียบร้อย",
           text: "",
           icon: "success",
           button: "ตกลง",
@@ -205,6 +229,10 @@
           console.log(data.responseJSON.errors);
         }
       });
+      } else {
+        
+      }
+    });
     });
   });
